Crate brainfuck

Simple brainfuck interpreter in Rust.

The brainfuck language was created with the purpose of being a very minimal language which is very easy to write an interpreter for. This is one such interpreter. For more information on the brainfuck language start with the documentation of each instruction in the language, or some material online. Brainfuck itself is syntactically challenging for humans, but is really not all that complicated. +. for example increments the value in the first cell and outputs that value. The instructions that trip people up the most are the control flow contructs [ and ]. +++>,<[>+.<-] for example prints the 3 values after the input value. For more about control flow in brainfuck read the section on control flow.

Semantics and Portability

There are a few extensions and devivations from the common semantics of the language. The brainfuck language has a few areas that are undefined behavior. These undefined behaviors are given explicit semantics in this implmentation. Most brainfuck programs should work as expected in this implmentation. For more information on portabiliy of brainfuck programs read The Unofficial Constraints on Portable Brainfuck Implementations.

TODO: Explain the choices made which are not spelled out for the language.

Structs

Interpreter
A brainfuck interpreter, with the needed state for execution.

Program

Enums

Error
A general error type for problems inside of the interpreter.

Instruction
An executable instruction in the language.