Struct brainfuck::Interpreter

pub struct Interpreter<'a> {
    // some fields omitted
}

A brainfuck interpreter, with the needed state for execution.

For more information about the brainfuck language in general see the top level documentation for this crate. The program code is stored as a string, and can be any size. The tape is an array of 30,000 unsigned bytes. This is derived from the original description of the language.

Brainfuck programs traditionally read from STDIN and STDOUT, but to make things a bit more general we allow reading and writing to arbitrary readers and writers.

Methods

impl<'a> Interpreter<'a>

fn new() -> Interpreter<'a>

Return a new interpreter with the given code, reader, and writter.

Interpreters are relatively large, so avoid too many calls to this function.

fn load(&mut self, program: Program) -> &mut Self

Load a program for the interpreter to run.

fn read_from<R: Read>(&mut self, reader: &'a mut R) -> &mut Self

fn write_to<W: Write>(&mut self, writer: &'a mut W) -> &mut Self

fn run(&mut self) -> Result<(), Error>

Run the interpreter.

fn run_with_callback<F>(&mut self, hook: F) -> Result<(), Error> where F: FnMut(&mut Self, &Instruction)

Run the interpreter with a callback hook.