Struct brainfuck::Interpreter

pub struct Interpreter {
    // some fields omitted
}

A brainfuck interpreter, with the needed state for execution.

For more information about the brainfuck language in general see the top level documentation for this crate. The program code is stored as a string, and can be any size. The tape is an array of 30,000 unsigned bytes. This is derived from the original description of the language.

Methods

impl Interpreter

fn load<P: AsRef<Path>>(path: P) -> Result<Interpreter, Error>

Create a new interpreter from a file.

Loads the given file from the path, and creates a new interpreter. The code of the interpreter is the contents of the file. Every cell of the tape is initialized to 0. Both the pointer and the program counter are set to 0. If the file fails to load then this function returns an error.

Examples

use brainfuck::Interpreter;

Interpreter::load("foo.bf");

fn run(&mut self)

Run the interpreter.

Examples

use brainfuck::Interpreter;

Interpreter::load("fixtures/hello.bf").unwrap().run();

fn run_with_callback<F>(&mut self, hook: F) where F: FnMut(&Instruction)

Run the interpreter with a callback hook.

Examples

use brainfuck::Interpreter;

let mut interp = Interpreter::load("fixtures/hello.bf").unwrap();
interp.run_with_callback(|i| {
    println!("Stepped: {}", i);
});

fn step(&mut self) -> Option<Instruction>

Step the interpreter one instruction.

This function returns None when there are no more steps to make in the code from the current value of the program counter.

Examples

use brainfuck::{Interpreter, Instruction};

let mut interp = Interpreter::load("fixtures/hello.bf").unwrap();
assert!(interp.step().unwrap() == Instruction::SkipForward);