Struct brainfuck::Interpreter

pub struct Interpreter<'a> {
    // some fields omitted
}

A brainfuck interpreter, with the needed state for execution.

For more information about the brainfuck language in general see the top level documentation for this crate. The program code is stored as a string, and can be any size. The tape is an array of 30,000 unsigned bytes. This is derived from the original description of the language.

Brainfuck programs traditionally read from STDIN and STDOUT, but to make things a bit more general we allow reading and writing to arbitrary readers and writers.

Examples

Print 1 to STDOUT.

use std::io;
use brainfuck::Interpreter;

let mut reader = io::stdin();
let mut writer = io::stdout();
Interpreter::new("+.", &mut reader, &mut writer).unwrap().run();

Methods

impl<'a> Interpreter<'a>

fn new<C: Into<String>, R: Read, W: Write>(code: C, input: &'a mut R, output: &'a mut W) -> Result<Interpreter<'a>, Error>

Return a new interpreter with the given code, reader, and writter.

Interpreters are relatively large, so avoid too many calls to this function.

Examples

use brainfuck::Interpreter;

let mut reader = &[][..];
let mut writer = Vec::<u8>::new();
Interpreter::new("+.", &mut reader, &mut writer).unwrap().run();

fn from_file<P: AsRef<Path>, R: Read, W: Write>(path: P, input: &'a mut R, output: &'a mut W) -> Result<Interpreter<'a>, Error>

Create a new interpreter from a file.

Loads the given file from the path, and creates a new interpreter. The code of the interpreter is the contents of the file. Every cell of the tape is initialized to 0. Both the pointer and the program counter are set to 0. If the file fails to load then this function returns an error.

Examples

use brainfuck::Interpreter;

let mut reader = &[][..];
let mut writer = Vec::<u8>::new();
let interp = Interpreter::from_file("fixtures/hello.b", &mut reader, &mut writer);

fn run(&mut self)

Run the interpreter.

Examples

use brainfuck::Interpreter;

let mut reader = &[][..];
let mut writer = Vec::<u8>::new();
Interpreter::from_file("fixtures/hello.b", &mut reader, &mut writer).unwrap().run();

fn run_with_callback<F>(&mut self, hook: F) where F: FnMut(&mut Self, &Instruction)

Run the interpreter with a callback hook.

Examples

use brainfuck::Interpreter;

let mut reader = &[][..];
let mut writer = Vec::<u8>::new();
let mut interp = Interpreter::from_file("fixtures/hello.b", &mut reader, &mut writer).unwrap();
interp.run_with_callback(|interp, inst| {
    println!("Stepped: {}", inst);
});

fn step(&mut self) -> Option<Result<Instruction, Error>>

Step the interpreter one instruction.

This function returns None when there are no more steps to make in the code from the current value of the program counter.

Examples

use brainfuck::{Interpreter, Instruction};

let mut reader = &[][..];
let mut writer = Vec::<u8>::new();
let mut interp = Interpreter::from_file("fixtures/hello.b", &mut reader, &mut writer).unwrap();
assert!(interp.step().unwrap().unwrap() == Instruction::SkipForward);